requirements.txt
gradio==5.22.0
google-cloud-logging==3.11.4
google-cloud-core==2.4.3
google-cloud-documentai==3.3.0
google-cloud-documentai-toolbox==0.14.2a0
PyPDF2==3.0.1
langchain-community==0.3.20
langchain-google-vertexai==2.0.16
tokenizers==0.21.1